test(client-python): prepare pr
