--- packages/client-python/tests/test_connection_uri_normalization.py
@@ -1,41 +1,0 @@
-# MIT License
-#
-# Copyright (c) 2026 Aparavi Software AG
-#
-# Permission is hereby granted, free of charge, to any person obtaining a copy
-# of this software and associated documentation files (the "Software"), to deal
-# in the Software without restriction, including without limitation the rights
-# to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-# copies of the Software, and to permit persons to whom the Software is
-# furnished to do so, subject to the following conditions:
-#
-# The above copyright notice and this permission notice shall be included in all
-# copies or substantial portions of the Software.
-#
-# THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-# IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-# FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-# AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-# LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-# OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-# SOFTWARE.
-
-"""Regression tests for websocket URI normalization."""
-
-import pytest
-
-from rocketride.mixins.connection import ConnectionMixin
-
-
-@pytest.mark.parametrize(
-    ('input_uri', 'expected_uri'),
-    [
-        ('wss://cloud.rocketride.ai', 'wss://cloud.rocketride.ai/task/service'),
-        ('https://cloud.rocketride.ai', 'wss://cloud.rocketride.ai/task/service'),
-        ('ws://localhost:5565', 'ws://localhost:5565/task/service'),
-        ('http://localhost:5565', 'ws://localhost:5565/task/service'),
-    ],
-)
-def test_get_websocket_uri_normalization(input_uri: str, expected_uri: str) -> None:
-    """Verify websocket URI normalization preserves secure and non-secure schemes."""
-    assert ConnectionMixin._get_websocket_uri(input_uri) == expected_uri
